data-transferring/requirements.txt
boto3
pandas
psycopg2-binary
python-dotenv